CP/DP Split: Add agent/nginx container and deployment (#2958)
Updating the nginx docker containers to build and include agent. Once agent is officially released, we can use the published binary instead of building.

Added a temporary nginx deployment to the helm chart to deploy a standalone nginx pod.

Added the basic gRPC server and agent API implementation to allow for the agent pod to connect to the control plane without errors.