Fix port binding with reduced privileges (#3574) (#3590)
Problem: The nginx deployment was using extra privileges in order to bind to privileged ports (<1024). This included `allowPrivilegeEscalation` and `NET_BIND_SERVICE`. Sometimes this could cause issues in some secure environments.

Solution: Remove these extra privileges and take advantage of `sysctls` to lower the allowed port range for the pod its defined on.