fix: remove patch label/ annotation cross contamination (#3754) (#3766)
Problem: Patching the Service object labels and annotations results in
the Deployment also getting these labels and annotations, and vice
versa, which is unexpected behavior.

Solution: Create separate copies of the base label and annotation maps
for each object. Since maps are passed by reference in Go, modifying the
labels for one object was affecting the other by updating the shared
reference.

Co-authored-by: Ciara Stacke <18287516+ciarams87@users.noreply.github.com>