Fix empty pod owner name when connecting to agent
Problem: Saw an issue where nginx failed to connect to the control plane due to an empty hostname and not being able to find the pod name.

Solution: If ContainerInfo provided by agent is empty, fall back to HostInfo. This should cover all bases to ensure we get the hostname.