Truncate long resource names when provisioning
Problem: When provisioning the NGINX Service, if the Gateway name was very long, it could result in a failure to create the Service. A Service name must be 63 characters or less. This also could affect the shadow Service that we create for InferencePools if the InferencePool name is long.

Solution: If the name of the resource that we create is going to be over the max, create a hash of the name and truncate to ensure that it is less than the max. The suffix is still included, since we rely on that for certain resources. Just the prefix is shortened as necessary to include the hash.