Update dockerfile packages
Since the OSS docker image uses an nginx image as the base, the alpine base is delayed in getting updates when there are CVEs and such.

This updates the apk packages to ensure that everything is always up to date for that distribution. Also removed unnecessary library from UBI build.