Add doc describing provisioning restrictions
A common problem users have is that they can't directly edit the NGINX resources that we provision, and we direct them to our docs on how to set the fields they want.

I'm adding this document here to describe why this restriction is in place. This can help developers on the team understand it, to help explain it if needed, and can also be used as a reference to any users who may be curious.